requirements.txt
anyio==4.9.0
asgiref==3.8.1
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.6.15
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.2.3
django-environ==0.12.0
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
keyring==25.6.0
more-itertools==10.7.0
msgpack==1.1.1
packaging==25.0
pbs-installer==2025.6.30
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
psycopg2-binary==2.9.10
pyproject_hooks==1.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
whitenoise==6.9.0
zipp==3.23.0
zstandard==0.23.0
